+++ Tests/Get-CloudPCConnectivityHistory.Tests.ps1
@@ -11,6 +11,7 @@ AfterAll {
 Describe 'Get-CloudPCConnectivityHistory' {
 
     It 'calls the Cloud PC connectivity history endpoint and normalizes events' {
+        Mock -ModuleName WindowsCloudPC Connect-CloudPC { }
         Mock -ModuleName WindowsCloudPC Invoke-GraphPaged -MockWith {
             @(
                 @{
@@ -24,9 +25,10 @@ Describe 'Get-CloudPCConnectivityHistory' {
             )
         }
 
-        $r = & (Get-Module WindowsCloudPC) { Get-CloudPCConnectivityHistory -CloudPcId 'cloud pc/id' }
+        $r = Get-CloudPCConnectivityHistory -CloudPcId 'cloud pc/id'
 
         $r.PSObject.TypeNames | Should -Contain 'WindowsCloudPC.CloudPCConnectivityEvent'
+        $r.CloudPcId | Should -Be 'cloud pc/id'
         $r.ActivityId | Should -Be 'activity-1'
         $r.EventDateTime | Should -BeOfType [datetime]
         Should -Invoke -ModuleName WindowsCloudPC Invoke-GraphPaged -Times 1 -Exactly -ParameterFilter {
@@ -34,10 +36,38 @@ Describe 'Get-CloudPCConnectivityHistory' {
         }
     }
 
+    It 'accepts Cloud PC objects from the pipeline' {
+        Mock -ModuleName WindowsCloudPC Connect-CloudPC { }
+        Mock -ModuleName WindowsCloudPC Invoke-GraphPaged -MockWith {
+            @(
+                @{
+                    activityId    = 'activity-2'
+                    eventDateTime = '2026-06-17T19:10:53Z'
+                    eventType     = 'userConnection'
+                    eventName     = 'Connection Finished'
+                    eventResult   = 'success'
+                    message       = ''
+                }
+            )
+        }
+
+        $pc = [pscustomobject]@{
+            PSTypeName = 'WindowsCloudPC.CloudPC'
+            Id         = 'cpc-1'
+            Name       = 'CPC-ONE'
+        }
+
+        $r = $pc | Get-CloudPCConnectivityHistory
+
+        $r.CloudPcId | Should -Be 'cpc-1'
+        $r.CloudPcName | Should -Be 'CPC-ONE'
+    }
+
     It 'returns null when Graph throws' {
+        Mock -ModuleName WindowsCloudPC Connect-CloudPC { }
         Mock -ModuleName WindowsCloudPC Invoke-GraphPaged -MockWith { throw 'boom' }
 
-        $r = & (Get-Module WindowsCloudPC) { Get-CloudPCConnectivityHistory -CloudPcId 'bad' }
+        $r = Get-CloudPCConnectivityHistory -CloudPcId 'bad'
 
         $r | Should -BeNullOrEmpty
     }