+++ Tests/WindowsCloudPC.Tests.ps1
@@ -4,6 +4,7 @@ BeforeDiscovery {
         'Export-CloudPCProvisioningPolicy',
         'Get-CloudPC',
         'Get-CloudPCByProvisioningPolicy',
+        'Get-CloudPCConnectivityHistory',
         'Get-CloudPCLaunchDetail',
         'Get-CloudPCLicensingAllotment',
         'Get-CloudPCMaintenanceWindow',